+++ RealmSwift/Realm.swift
@@ -208,6 +208,18 @@ public final class Realm {
         rlmRealm.cancelWriteTransaction()
     }
 
+    /**
+    Indicates if this Realm is currently in a write transaction.
+
+    :warning: Wrapping mutating operations in a write transaction if this property returns `false`
+              may cause a large number of write transactions to be created, which could negatively
+              impact Realm's performance. Always prefer performing multiple mutations in a single
+              transaction when possible.
+    */
+    public var inWriteTransaction: Bool {
+        return rlmRealm.inWriteTransaction
+    }
+
     // MARK: Adding and Creating objects
 
     /**